Rewrite tgen plots to use pandas and seaborn.
Previous OnionPerf visualizations were written using PyLab, which is a
relatively low-level visualization library based on matplotlib. PyLab
is the same thing as PyPlot but with some extra shortcuts. However,
using PyLab is discouraged now.

This commit switches all tgen plots to use seaborn, which is a
high-level data visualization library also based on matplotlib. This
commit also uses pandas for underlying data structures in order to
separate data tidying and data visualization.

As a side effect of using pandas, graphed data is now exported to a
.csv file using the same naming scheme as the produced tgen .pdf file.

Implements #33258.