Reduce timeout and stallout values.
The idea is to avoid overlapping measurements. With a new measurement
starting every 5:00 minutes, we time out after 4:30 minutes to make
"room" for the next measurement to start.

We also don't use the stallout feature anymore by setting the value to
0 seconds. The stallout would have failed a measurement after not
receiving new bytes for the given number of seconds. But that's also
not what we want to measure.

Implements #34024.