Work around Networkx's API change.
Turns out that Networkx renamed the node attribute to nodes some time
between version 2.2.1 and 2.4. This workaround avoids accessing that
attribute to add another key/value pair to an existing node and
instead uses two different key/value pair sets when adding the node in
the first place. Not pretty, but works around the issue.

Fixes #34298.