Make README.md more useful.
This commit attempts to make README.md the primary, comprehensive
documentation for OnionPerf. There will be more documentation on the
wiki for things that change faster than we want to update a file in
Git or that are too specific for most users. But this README.md is
supposed to cover 90% of the questions that new users would have.

Implements #40001.